% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{assignC}
\alias{assignC}
\title{Concatenate to a variable by name}
\usage{
assignC(t, a, env = parent.frame())
}
\arguments{
\item{t}{A string containing the name of the variable to be modified.}

\item{a}{The value(s) to be concatenated.}

\item{env}{The environment in which to look for the variable (default: parent.frame()).}
}
\value{
The modified variable (invisibly).
}
\description{
This function concatenates a value to a variable specified by its name.
}
\examples{
x <- c(1, 2, 3)
assignC("x", c(4, 5))
print(x)  # Output: 1 2 3 4 5

}
