% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{assignMinus}
\alias{assignMinus}
\title{Subtract from a variable by name}
\usage{
assignMinus(t, m, env = parent.frame())
}
\arguments{
\item{t}{A string containing the name of the variable to be modified.}

\item{m}{The value to be subtracted.}

\item{env}{The environment in which to look for the variable (default: parent.frame()).}
}
\value{
The modified variable (invisibly).
}
\description{
This function subtracts a value from a variable specified by its name.
}
\examples{
x <- 10
assignMinus("x", 4)
print(x)  # Output: 6

}
