% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{assignPlus}
\alias{assignPlus}
\title{Add to a variable by name}
\usage{
assignPlus(t, s, env = parent.frame())
}
\arguments{
\item{t}{A string containing the name of the variable to be modified.}

\item{s}{The value to be added.}

\item{env}{The environment in which to look for the variable (default: parent.frame()).}
}
\value{
The modified variable (invisibly).
}
\description{
This function adds a value to a variable specified by its name.
}
\examples{
x <- 5
assignPlus("x", 3)
print(x)  # Output: 8

}
