% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{assignSetdiff}
\alias{assignSetdiff}
\title{Set difference with a variable by name}
\usage{
assignSetdiff(t, a, env = parent.frame())
}
\arguments{
\item{t}{A string containing the name of the variable to be modified.}

\item{a}{The value(s) to be removed from the variable.}

\item{env}{The environment in which to look for the variable (default: parent.frame()).}
}
\value{
The modified variable (invisibly).
}
\description{
This function performs a set difference of a variable specified by its name with the given values.
}
\examples{
x <- c(1, 2, 3, 4, 5)
assignSetdiff("x", c(3, 4))
print(x)  # Output: 1 2 5

}
