% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{assignUnion}
\alias{assignUnion}
\title{Union with a variable by name}
\usage{
assignUnion(t, a, env = parent.frame())
}
\arguments{
\item{t}{A string containing the name of the variable to be modified.}

\item{a}{The value(s) to be unioned.}

\item{env}{The environment in which to look for the variable (default: parent.frame()).}
}
\value{
The modified variable (invisibly).
}
\description{
This function performs a union of a value with a variable specified by its name.
}
\examples{
x <- c(1, 2, 3)
assignUnion("x", c(3, 4, 5))
print(x)  # Output: 1 2 3 4 5

}
