% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%c=\%}
\alias{\%c=\%}
\title{Concatenate to a variable in-place}
\usage{
t \%c=\% a
}
\arguments{
\item{t}{The variable to be modified.}

\item{a}{The value(s) to be concatenated.}
}
\value{
The modified variable (invisibly).
}
\description{
This operator concatenates the right-hand side to the left-hand side variable in-place.
}
\examples{
x <- c(1, 2, 3)
x \%c=\% c(4, 5)
print(x)  # Output: 1 2 3 4 5

}
