% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%-=\%}
\alias{\%-=\%}
\title{Subtract from a variable in-place}
\usage{
t \%-=\% m
}
\arguments{
\item{t}{The variable to be modified.}

\item{m}{The value to be subtracted.}
}
\value{
The modified variable (invisibly).
}
\description{
This operator subtracts the right-hand side from the left-hand side variable in-place.
}
\examples{
x <- 10
x \%-=\% 4
print(x)  # Output: 6

}
