% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%+=\%}
\alias{\%+=\%}
\title{Add to a variable in-place}
\usage{
t \%+=\% s
}
\arguments{
\item{t}{The variable to be modified.}

\item{s}{The value to be added.}
}
\value{
The modified variable (invisibly).
}
\description{
This operator adds the right-hand side to the left-hand side variable in-place.
}
\examples{
x <- 5
x \%+=\% 3
print(x)  # Output: 8

}
