% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%setdiff=\%}
\alias{\%setdiff=\%}
\title{Set difference with a variable in-place}
\usage{
t \%setdiff=\% a
}
\arguments{
\item{t}{The variable to be modified.}

\item{a}{The value(s) to be removed from t.}
}
\value{
The modified variable (invisibly).
}
\description{
This operator performs a set difference of the left-hand side variable with the right-hand side in-place.
}
\examples{
x <- c(1, 2, 3, 4, 5)
x \%setdiff=\% c(3, 4)
print(x)  # Output: 1 2 5

}
