% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_ops.R
\name{\%union=\%}
\alias{\%union=\%}
\title{Union with a variable in-place}
\usage{
t \%union=\% a
}
\arguments{
\item{t}{The variable to be modified.}

\item{a}{The value(s) to be unioned.}
}
\value{
The modified variable (invisibly).
}
\description{
This operator performs a union of the right-hand side with the left-hand side variable in-place.
}
\examples{
x <- c(1, 2, 3)
x \%union=\% c(3, 4, 5)
print(x)  # Output: 1 2 3 4 5

}
