% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{inplaceops-package}
\alias{inplaceops}
\alias{inplaceops-package}
\title{inplaceops: Inplace Operators for R}
\description{
Inplace operators for R. This does operations like `x += 1` by writing `1 %+=% x`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PROGR-2020/r_package_demo2}
  \item Report bugs at \url{https://github.com/PROGR-2020/r_package_demo2/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{things.developer@mb706.com}

}
